extract extra-data component, refactor selectors
