+++ examples/todomvc/src/selectors/index.js
@@ -1,10 +1,9 @@
 import { createSelector } from 'reselect'
 import { SHOW_ALL, SHOW_COMPLETED, SHOW_ACTIVE } from '../constants/TodoFilters'
-import _ from 'lodash'
+import { createWcSelector } from '../wildcard/selectors'
 
 const getVisibilityFilter = state => state.visibilityFilter
 const getTodos = state => state.todos
-const getWcConfig = state => state.__wildcard__
 
 export const getVisibleTodos = createSelector(
   [getVisibilityFilter, getTodos],
@@ -22,6 +21,8 @@ export const getVisibleTodos = createSelector(
   }
 )
 
+export const getSortedTodos = createWcSelector(getVisibleTodos);
+
 export const getCompletedTodoCount = createSelector(
   [getTodos],
   todos => (
@@ -30,15 +31,4 @@ export const getCompletedTodoCount = createSelector(
       0
     )
   )
-)
-
-export const getSortedTodos = createSelector(
-  [getVisibleTodos, getWcConfig],
-  (visibleTodos, wcConfig) => {
-    const sortConfig = wcConfig.sort
-    if (!sortConfig || !sortConfig.direction === "NONE") { return visibleTodos; }
-    let sorted = _.sortBy(visibleTodos, t => t[sortConfig.column])
-    if (sortConfig.direction === "DESC") { sorted = sorted.reverse() }
-    return sorted;
-  }
 )
