+++ examples/todomvc/src/wildcard/configReducer.js
